planner-up/requirements.txt
grpcio==1.37.0
grpcio-tools==1.37.0
jupyter==1.0.0
unified-planning==0.4.2
unified-planning[fast-downward]
